+++ src/read/decoder.rs
@@ -20,7 +20,7 @@ const DECODED_CHUNK_SIZE: usize = 3;
 /// let mut wrapped_reader = Cursor::new(b"YXNkZg==");
 /// let mut decoder = base64::read::DecoderReader::new(
 ///     &mut wrapped_reader,
-///     &base64::engine::DEFAULT_ENGINE);
+///     &base64::engine::STANDARD);
 ///
 /// // handle errors as you normally would
 /// let mut result = Vec::new();